backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
pyodbc==5.2.0
pydantic-settings==2.10.1
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt<4.1
httpx==0.28.1
python-multipart==0.0.20
email-validator==2.2.0
reportlab==4.0.9
openpyxl==3.1.5
python-dotenv==1.0.0
websockets==13.0
matplotlib==3.9.2
